+++ src/main/java/moze_intel/projecte/PECore.java
@@ -106,7 +106,7 @@ import org.slf4j.Logger;
 public class PECore {
 
 	public static final String MODID = ProjectEAPI.PROJECTE_MODID;
-	public static final String MODNAME = "ProjectE";
+	public static final String MODNAME = "ProjectEE";
 	public static final GameProfile FAKEPLAYER_GAMEPROFILE = new GameProfile(UUID.fromString("590e39c7-9fb6-471b-a4c2-c0e539b2423d"), "[" + MODNAME + "]");
 	public static final Logger LOGGER = LogUtils.getLogger();